ДЗ модуль 5 - разбиение на модули
