+++ index.html
@@ -234,7 +234,6 @@
       </section>
     </template>
 
-  <script src="./js/functions.js"></script>
-  <script src="./js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 </html>
